+++ test/unit/test_messages.py
@@ -51,12 +51,30 @@ def test_slack_exception(mock_slack, mock_logger, mock_messages, mock_token, moc
 
 
 def test_prepare_pulls_message(mock_pull_request, mock_user, mock_repo):
+    """Tests that we build user strings and messages correctly when present."""
     reviewer = MagicMock()
-    reviewer.login = mock_user
+    reviewer.login = 'reviewer'
     reviewer.html_url = f'https://github.com/{mock_user}'
     reviewers = [reviewer]
 
-    slack_message, discord_message = Message.prepare_pulls_message(mock_pull_request, reviewers, reviewers, reviewers)
+    approved_reviewer = MagicMock()
+    approved_reviewer.login = 'approved_reviewer'
+    approved_reviewer.html_url = f'https://github.com/{mock_user}'
+    approved_reviewers = [approved_reviewer]
+
+    requested_changes_reviewer = MagicMock()
+    requested_changes_reviewer.login = 'requested_changes_reviewer'
+    requested_changes_reviewer.html_url = f'https://github.com/{mock_user}'
+    requested_changes_reviewers = [requested_changes_reviewer]
+
+    dismissed_reviewer = MagicMock()
+    dismissed_reviewer.login = 'dismissed_reviewer'
+    dismissed_reviewer.html_url = f'https://github.com/{mock_user}'
+    dismissed_reviewers = [dismissed_reviewer]
+
+    slack_message, discord_message = Message.prepare_pulls_message(
+        mock_pull_request, reviewers, approved_reviewers, requested_changes_reviewers, dismissed_reviewers
+    )
 
     # Slack message
     assert 'Pull Request' in slack_message
@@ -69,12 +87,35 @@ def test_prepare_pulls_message(mock_pull_request, mock_user, mock_repo):
     assert f'{mock_pull_request.title} (<{mock_pull_request.html_url}>)' in discord_message
 
 
-def test_prepare_pulls_message_no_assignee(mock_pull_request):
-    mock_pull_request.requested_reviewers = []
-    slack_message, discord_message = Message.prepare_pulls_message(mock_pull_request, [], [], [])
+def test_prepare_pulls_message_same_reviewer(mock_pull_request, mock_user, mock_repo):
+    """Ensures that when a user has requested changes, been dismissed, then approved, we filter those correctly."""
+    reviewer = MagicMock()
+    reviewer.login = 'reviewer'
+    reviewer.html_url = f'https://github.com/{mock_user}'
+    reviewers = [reviewer]
 
-    assert 'Reviewers:* NA' in slack_message
-    assert 'Reviewers:* NA' in discord_message
+    slack_message, discord_message = Message.prepare_pulls_message(
+        mock_pull_request, reviewers, reviewers, reviewers, reviewers
+    )
+
+    # Slack message
+    assert 'Pull Request' in slack_message
+    assert f'{reviewer.html_url}|{reviewer.login}' in slack_message
+    assert f'{mock_pull_request.html_url}|{mock_pull_request.title}' in slack_message
+
+    # Discord message
+    assert 'Pull Request' in discord_message
+    assert f'{reviewer.login} (<{reviewer.html_url}>)' in discord_message
+    assert f'{mock_pull_request.title} (<{mock_pull_request.html_url}>)' in discord_message
+
+
+def test_prepare_pulls_message_no_reviewers(mock_pull_request):
+    """Tests that no user strings are generated when there are no reviewers."""
+    mock_pull_request.requested_reviewers = []
+    slack_message, discord_message = Message.prepare_pulls_message(mock_pull_request, [], [], [], [])
+
+    assert '*Reviewers:* NA' in slack_message
+    assert '*Reviewers:* NA' in discord_message
 
 
 def test_prepare_issues_message(mock_issue, mock_user, mock_repo):